feat: 웹 UI에 말을 시작 배치로 세우고 점수 표기하기
