+++ src/main/java/chess/ChessApplication.java
@@ -1,15 +1,16 @@
 package chess;
 
 import chess.controller.ChessController;
+import chess.controller.ConsoleChessController;
 import chess.view.OutputView;
 
 public class ChessApplication {
-	public static void main(String[] args) {
-		ChessController chessController = new ChessController();
-		try {
-			chessController.run();
-		} catch (IllegalArgumentException e) {
-			OutputView.printErrorMessage(e.getMessage());
-		}
-	}
+    public static void main(String[] args) {
+        ChessController consoleChessController = new ConsoleChessController();
+        try {
+            consoleChessController.run();
+        } catch (IllegalArgumentException e) {
+            OutputView.printErrorMessage(e.getMessage());
+        }
+    }
 }
